+++ bin/update_config.js
@@ -33,14 +33,27 @@ function update_config(config) {
   Object.keys(globalConfig.Miniserver).forEach(key => {
     if (!config.miniserver[key]) {
       let ms = {
-        "mqtt_topic_ms": "loxone",
-        "publish_structure": false,
-        "publish_states": false,
-        "subscribe": false,
-        "mqtt_topic_app": "loxberry/app",
-        "icon_path": "/assets/icons/svg"
+        mqtt_topic_ms: "loxone",
+        publish_structure: false,
+        publish_states: false,
+        subscribe: false,
+        other_user: false,
+        user: null,
+        pass: null,
+        mqtt_topic_app: "loxberry/app"
       };
       config.miniserver[key] = ms;
+    }
+    else { // exists, check/update each individual item
+      let ms = config.miniserver[key];
+      if (!ms.mqtt_topic_ms) ms.mqtt_topic_ms = "loxone";
+      if (!ms.publish_structure) ms.publish_structure = false;
+      if (!ms.publish_states) ms.publish_states = false;
+      if (!ms.subscribe) ms.subscribe = false;
+      if (!ms.other_user) ms.other_user = false;
+      if (!ms.user) ms.user = null;
+      if (!ms.pass) ms.pass = null;
+      if (!ms.mqtt_topic_app) ms.mqtt_topic_app = "loxberry/app";
     }
   });